refactor: refactor Search component from class to functional component
